Further refactoring of MonthlyReport stuff
* refactored journey warning calculation to strategy pattern
* renamed some calculation classes so they are named uniformly
