+++ mega-zep-backend/src/main/java/com/gepardec/mega/service/impl/monthlyreport/MonthlyReportServiceImpl.java
@@ -6,6 +6,7 @@ import com.gepardec.mega.domain.model.monthlyreport.MonthlyReport;
 import com.gepardec.mega.domain.model.monthlyreport.ProjectTimeEntry;
 import com.gepardec.mega.domain.model.monthlyreport.TimeWarning;
 import com.gepardec.mega.service.api.monthlyreport.MonthlyReportService;
+import com.gepardec.mega.service.impl.monthlyreport.calculation.WarningCalculator;
 import com.gepardec.mega.zep.ZepService;
 
 import javax.enterprise.context.RequestScoped;
